+++ test/authentication/ticket_authentication_test.dart
@@ -1,3 +1,5 @@
+import 'dart:async';
+
 import 'package:connectanum/src/authentication/ticket_authentication.dart';
 import 'package:connectanum/src/message/challenge.dart';
 import 'package:test/test.dart';
@@ -11,6 +13,18 @@ void main() {
       expect(authMethod.getName(), equals('ticket'));
       final authenticate = await authMethod.challenge(Extra());
       expect(authenticate.signature, equals(secret));
+    });
+    test('on challenge event', () async {
+      final authMethod = TicketAuthentication("test");
+      final completer = Completer<Extra>();
+      authMethod.onChallenge.listen((event) {
+        completer.complete(event);
+      },);
+      var extra = Extra(challenge: "challenge", channelBinding: null);
+      authMethod.challenge(extra);
+      var receivedExtra = await completer.future;
+      expect(receivedExtra, isNotNull);
+      expect(receivedExtra.challenge, equals("challenge"));
     });
   });
 }